server/requirements.txt
fastapi
uvicorn
supabase
fastembed
google-genai
gitpython
python-dotenv
prisma
pydantic
tree-sitter>=0.21.0
tree-sitter-language-pack>=0.1.0
sentence-transformers>=2.2.0
aisuite
tavily-python
wikipedia
pdfminer.six
PyMuPDF
fitz
